feat(PN-1248): wip on modal functionality
