+++ packages/pn-personafisica-webapp/public/locales/it/recapiti.json
@@ -128,7 +128,7 @@
     "sms-add-more-button": "Aggiungi altro numero",
     "sms-add-more-caption": "Vuoi che le notifiche inviate da alcuni enti siano indirizzate a un altro numero?",
     "email-add-more-button": "Aggiungi altra email",
-    "email-add-more-caption": "Vuoi che le notifiche inviate da alcuni enti siano indirizzate a un'altro email?",
+    "email-add-more-caption": "Vuoi che le notifiche inviate da alcuni enti siano indirizzate a un'altra email?",
     "pec-add-more-button": "Aggiungi altra PEC",
     "pec-add-more-caption": "Vuoi che le notifiche inviate da alcuni enti siano indirizzate a un'altra PEC?",
     "sender-list": "Per le notifiche inviate da:",
