+++ packages/pn-personafisica-webapp/src/components/Contacts/AddSpecialContactDialog.tsx
@@ -1,23 +1,32 @@
-import { useState } from 'react';
+import { ChangeEvent, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { Button, DialogTitle, Stack, TextField, Typography } from '@mui/material';
-import { PnAutocomplete, PnDialog, PnDialogActions, PnDialogContent } from '@pagopa-pn/pn-commons';
+import { Button, DialogTitle, MenuItem, Stack, TextField, Typography } from '@mui/material';
+import {
+  PnAutocomplete,
+  PnDialog,
+  PnDialogActions,
+  PnDialogContent,
+  searchStringLimitReachedText,
+} from '@pagopa-pn/pn-commons';
 
+import { DigitalAddress } from '../../models/contacts';
 import { Party } from '../../models/party';
 import { useAppSelector } from '../../redux/hooks';
 import { RootState } from '../../redux/store';
+import DropDownPartyMenuItem from '../Party/DropDownParty';
 
 type Props = {
   open: boolean;
   senderId?: string;
   handleClose: () => void;
   formik: any;
-  handleChangeTouched: () => {};
   handleConfirm: () => void;
-  renderOption: any;
-  senderChangeHandler: any;
-  entitySearchLabel: any;
+  digitalAddresses: Array<DigitalAddress>;
+};
+
+type Addresses = {
+  [senderId: string]: Array<DigitalAddress>;
 };
 
 const AddSpecialContactDialog: React.FC<Props> = ({
@@ -25,16 +34,64 @@ const AddSpecialContactDialog: React.FC<Props> = ({
   senderId = 'default',
   handleClose,
   formik,
-  handleChangeTouched,
   handleConfirm,
-  renderOption,
-  senderChangeHandler,
-  entitySearchLabel,
+  digitalAddresses,
 }) => {
   const { t } = useTranslation(['common', 'recapiti']);
   const getOptionLabel = (option: Party) => option.name || '';
   const [senderInputValue, setSenderInputValue] = useState('');
+  const [alreadyExistsMessage, setAlreadyExistsMessage] = useState('');
+
   const parties = useAppSelector((state: RootState) => state.contactsState.parties);
+
+  const addresses: Addresses = digitalAddresses
+    .filter((a) => a.senderId !== 'default')
+    .reduce((obj, a) => {
+      if (!obj[a.senderId]) {
+        // eslint-disable-next-line functional/immutable-data
+        obj[a.senderId] = [];
+      }
+      // eslint-disable-next-line functional/immutable-data
+      obj[a.senderId].push(a);
+      return obj;
+    }, {} as Addresses);
+
+  const contactType = formik.values.addressType?.toLowerCase();
+
+  const senderChangeHandler = async (_: any, newValue: Party | null) => {
+    await formik.setFieldTouched('sender', true, false);
+    await formik.setFieldValue('sender', newValue);
+    setSenderInputValue(newValue?.name ?? '');
+    if (newValue && addresses[newValue.id]) {
+      const alreadyExists =
+        addresses[newValue.id].findIndex((a) => a.channelType === formik.values.addressType) > -1;
+      setAlreadyExistsMessage(
+        alreadyExists
+          ? t(`special-contacts.${contactType}-already-exists`, {
+              ns: 'recapiti',
+            })
+          : ''
+      );
+      return;
+    }
+    setAlreadyExistsMessage('');
+  };
+
+  // handling of search string for sender
+  const entitySearchLabel: string = `${t('special-contacts.add-sender', {
+    ns: 'recapiti',
+  })}${searchStringLimitReachedText(senderInputValue)}`;
+
+  const renderOption = (props: any, option: Party) => (
+    <MenuItem {...props} value={option.id} key={option.id}>
+      <DropDownPartyMenuItem name={option.name} />
+    </MenuItem>
+  );
+
+  const handleChangeTouched = async (e: ChangeEvent) => {
+    formik.handleChange(e);
+    await formik.setFieldTouched(e.target.id, true, false);
+  };
 
   return (
     <PnDialog open={open} onClose={handleClose} data-testid="cancelVerificationModal">
