+++ packages/pn-personafisica-webapp/src/components/Contacts/SpecialContacts.tsx
@@ -4,7 +4,7 @@ import { Trans, useTranslation } from 'react-i18next';
 import * as yup from 'yup';
 
 import VerifiedIcon from '@mui/icons-material/Verified';
-import { Box, Divider, MenuItem, Stack, Typography } from '@mui/material';
+import { Box, Divider, Stack, Typography } from '@mui/material';
 import {
   ApiErrorWrapper,
   AppResponse,
@@ -12,7 +12,6 @@ import {
   CodeModal,
   ErrorMessage,
   appStateActions,
-  searchStringLimitReachedText,
   useIsMobile,
 } from '@pagopa-pn/pn-commons';
 import { ButtonNaked } from '@pagopa/mui-italia';
@@ -24,7 +23,6 @@ import {
   DigitalAddress,
   SaveDigitalAddressParams,
 } from '../../models/contacts';
-import { Party } from '../../models/party';
 import {
   CONTACT_ACTIONS,
   createOrUpdateAddress,
@@ -40,7 +38,6 @@ import {
   pecValidationSchema,
   phoneValidationSchema,
 } from '../../utility/contacts.utility';
-import DropDownPartyMenuItem from '../Party/DropDownParty';
 import AddSpecialContactDialog from './AddSpecialContactDialog';
 import ExistingContactDialog from './ExistingContactDialog';
 import PecVerificationDialog from './PecVerificationDialog';
@@ -48,12 +45,6 @@ import PecVerificationDialog from './PecVerificationDialog';
 type Props = {
   digitalAddresses: Array<DigitalAddress>;
   addressType: ChannelType;
-  handleChangeTouched: any;
-  handleConfirm: any;
-};
-
-type Addresses = {
-  [senderId: string]: Array<DigitalAddress>;
 };
 
 type AddressTypeItem = {
@@ -68,17 +59,10 @@ enum ModalType {
   SPECIAL = 'special',
 }
 
-const SpecialContacts: React.FC<Props> = ({
-  digitalAddresses,
-  addressType,
-  handleChangeTouched,
-  handleConfirm,
-}) => {
+const SpecialContacts: React.FC<Props> = ({ digitalAddresses, addressType }) => {
   const { t } = useTranslation(['common', 'recapiti']);
   const isMobile = useIsMobile();
   const dispatch = useAppDispatch();
-  const [alreadyExistsMessage, setAlreadyExistsMessage] = useState('');
-  // const isMobile = useIsMobile();
   const [senderInputValue, setSenderInputValue] = useState('');
   const [modalOpen, setModalOpen] = useState<ModalType | null>(null);
   const codeModalRef =
@@ -90,18 +74,6 @@ const SpecialContacts: React.FC<Props> = ({
       id: a.channelType,
       value: t(`special-contacts.${a.channelType?.toLowerCase()}`, { ns: 'recapiti' }),
     }));
-
-  const addresses: Addresses = digitalAddresses
-    .filter((a) => a.senderId !== 'default')
-    .reduce((obj, a) => {
-      if (!obj[a.senderId]) {
-        // eslint-disable-next-line functional/immutable-data
-        obj[a.senderId] = [];
-      }
-      // eslint-disable-next-line functional/immutable-data
-      obj[a.senderId].push(a);
-      return obj;
-    }, {} as Addresses);
 
   const fetchAllActivatedParties = useCallback(() => {
     void dispatch(getAllActivatedParties({}));
@@ -160,36 +132,6 @@ const SpecialContacts: React.FC<Props> = ({
     formik.values.addressType === ChannelType.PEC ? 'legal-contacts' : 'courtesy-contacts';
   const contactType = formik.values.addressType?.toLowerCase();
 
-  const renderOption = (props: any, option: Party) => (
-    <MenuItem {...props} value={option.id} key={option.id}>
-      <DropDownPartyMenuItem name={option.name} />
-    </MenuItem>
-  );
-
-  // handling of search string for sender
-  const entitySearchLabel: string = `${t('special-contacts.sender', {
-    ns: 'recapiti',
-  })}${searchStringLimitReachedText(senderInputValue)}`;
-
-  const senderChangeHandler = async (_: any, newValue: Party | null) => {
-    await formik.setFieldTouched('sender', true, false);
-    await formik.setFieldValue('sender', newValue);
-    setSenderInputValue(newValue?.name ?? '');
-    if (newValue && addresses[newValue.id]) {
-      const alreadyExists =
-        addresses[newValue.id].findIndex((a) => a.channelType === formik.values.addressType) > -1;
-      setAlreadyExistsMessage(
-        alreadyExists
-          ? t(`special-contacts.${contactType}-already-exists`, {
-              ns: 'recapiti',
-            })
-          : ''
-      );
-      return;
-    }
-    setAlreadyExistsMessage('');
-  };
-
   const sendSuccessEvent = (type: ChannelType) => {
     const event =
       type === ChannelType.PEC
@@ -211,12 +153,15 @@ const SpecialContacts: React.FC<Props> = ({
       PFEventStrategyFactory.triggerEvent(event, formik.values.sender.id);
     }
 
+    console.log('siamo qua 1');
     const addressType =
       formik.values.addressType === ChannelType.PEC ? AddressType.LEGAL : AddressType.COURTESY;
     const value =
       formik.values.addressType === ChannelType.SMS
         ? internationalPhonePrefix + formik.values.s_value
         : formik.values.s_value;
+
+    console.log('value', value);
     const digitalAddressParams: SaveDigitalAddressParams = {
       addressType,
       senderId: formik.values.sender.id,
@@ -225,6 +170,7 @@ const SpecialContacts: React.FC<Props> = ({
       value,
       code: verificationCode,
     };
+    console.log('digitalAddressParams', digitalAddressParams);
 
     dispatch(createOrUpdateAddress(digitalAddressParams))
       .unwrap()
@@ -249,6 +195,7 @@ const SpecialContacts: React.FC<Props> = ({
               }),
             })
           );
+
           setModalOpen(null);
           // reset form
           formik.resetForm();
@@ -422,14 +369,11 @@ const SpecialContacts: React.FC<Props> = ({
         handleConfirm={() => setModalOpen(null)}
       />
       <AddSpecialContactDialog
-        open={true}
-        handleClose={() => {}}
+        open={modalOpen === ModalType.SPECIAL}
+        handleClose={() => setModalOpen(null)}
         formik={formik}
-        handleChangeTouched={handleChangeTouched}
-        handleConfirm={handleConfirm}
-        renderOption={renderOption}
-        senderChangeHandler={senderChangeHandler}
-        entitySearchLabel={entitySearchLabel}
+        handleConfirm={() => handleCodeVerification()}
+        digitalAddresses={digitalAddresses}
       />
       <Divider
         sx={{
